jenkins: Disable mutual auth
Main reason is to avoid failure on hubble flow validation if mTLS is
enabled. This could be just related to the test, just a note that GHA
jobs already provide required coverage. Also, this will help to avoid
the below issue in jenkins:

```
the namespace from the provided object "cilium-spire" does not match the namespace
"kube-system". You must pass '--namespace=cilium-spire' to perform this operation.
```

Signed-off-by: Tam Mach <tam.mach@cilium.io>