bpf: Make DEFINE_U*() macros use .data section for declarations
In the bpf package, where we process object files to load programs, we
skip rewritting .bss references for object files without a a .data
section. Working on a new feature, this caused some older kernels to
complain with "unrecognized bpf_ld_imm64 insn" (since they don't support
pseudo-register loads), some newer kernels with other errors apparently
related to the same missing references.

A first attempt at fixing the issue consisted in delaying the check for
the presence of the .data section, so that .bss references would be
inlined [0]. This would indeed fix the errors observed in the verifier
tests, but would instead break our BPF unit tests, where .bss is used to
store mock-up value that we want to handle in a specific way.

We need a long-term solution for dealing with the different sections in
a way that works for all set-ups; ideally, we want to be able to use
eBPF global variables everywhere we would need them, but this can't
happen before we drop support for older kernels.

In the meantime, let's macros that can cause declarations to add
variables to the .bss section, in prevision for the next commit. We
explicitely store these variables into .data instead.

[0] https://github.com/cilium/cilium/pull/25837/commits/50f8828510a743e87e18246ea9fe6c96ccaef5e5

Suggested-by: Timo Beckers <timo@isovalent.com>
Suggested-by: Dylan Reimerink <dylan.reimerink@isovalent.com>
Signed-off-by: Quentin Monnet <quentin@isovalent.com>